x64 Phase 2b (2/N): Utility / PathName / RegExp / cedtElement.cpp
Second chunk of the size_t-narrowing cleanup. Knocked another 80
warnings off (237 → 157 on Debug-KR|x64 clean rebuild) by walking
the four next-biggest .cpp files.

Patterns in this batch are the predictable x64-shaped ones —
explicit (INT) casts at the boundary between a 64-bit source
expression and a 32-bit destination variable, with the comment-
worthy ones noted below.

src/util/Utility.cpp (32 warnings)

  All instances of two patterns:
  - INT nLenN = strlen(...)   → INT nLenN = (INT)strlen(...)
  - nFound / nLocal = (pResult|pOffset) - pString [+ nResult]
                                → wrapped in (INT)( ... )
  Each function caps the input via "if (nFrom > nLen1) return -1"
  before doing pointer arithmetic, so the narrowing is bounded.

src/util/PathName.cpp (14 warnings)

  Six "INT n = strlen(p)" call sites — same single replacement as
  Utility.cpp. PathName.cpp has one extra C4996 (deprecated stricmp)
  in MatchFileFilter; that one's intentionally left for Phase 2c
  along with the rest of the deprecation cleanup.

src/util/RegExp.cpp (14 warnings)

  Spread of seven distinct narrowing sites in Henry Spencer's regex
  engine code, all of them "pointer difference fits in int because
  the input string is short" — line 502 stores a pointer offset
  into the regex bytecode as short (already 32K-bounded by the
  regex format itself), and the rest are matched-substring lengths
  into int. (short) and (int) casts respectively.

src/core/cedtElement.cpp (22 warnings)

  - One streamoff (CFile/fstream tellg) → UINT for a progress
    callback: (UINT)fin.tellg().
  - One CString constructor passed (fwd - beg) as length:
    CString(beg, (int)(fwd - beg)).
  - Three "INT nLength = wcslen(...)" / "INT nLength = strlen(...)"
    in the file-write path (one per encoding branch).
  - Two GetKeyNameText calls where the LPARAM scancode-shifted
    value goes into the GetKeyNameText LONG parameter — (LONG)
    cast at the call.
  - Four "nCount = m_lst*.GetCount()" in CMacroBuffer::StreamSave —
    same pattern as the inline accessors fixed in commit d0c3840,
    just spelled at the call sites this time.

Effect on Debug-KR|x64 clean rebuild after this commit:

  before this chunk : 237 warnings
  after  this chunk : 157 warnings  (down 80)

Cumulative on the branch:

  start of Phase 2  : 756 warnings (pointer-trunc not counted)
  end of  Phase 2b  : 157 warnings  (down 599 = ~79%)

Remaining ~150 warnings spread across roughly 20 .cpp files with
10 or fewer each (FileWndDirectory, cedtViewCommand, evaluate,
cedtAppConf, cedtView, RemoteFile, cedtDocEditAdv, cedtDocSearch,
FileTab, FileWndProject, cedtViewSearch, encode, ...) plus the
~19 non-narrowing warnings reserved for Phase 2c (C4996 / C4090 /
C4091 / C5033). All cedt_kr.exe x64 surfaces still build with
0 errors; smoke-test boot from build\x64\Release-KR\ unchanged.

+++ src/core/cedtElement.cpp
@@ -352,7 +352,7 @@ BOOL CDictionary::FileLoad(LPCTSTR lpszPathName, CALLBACK_FUNCTION fcnCallback)
 		}
 
 		// call a callback function if exist
-		if( ! (nCount % 100) && fcnCallback ) fcnCallback( fin.tellg() );
+		if( ! (nCount % 100) && fcnCallback ) fcnCallback( (UINT)fin.tellg() );
 	}
 
 	fin.close();
@@ -558,7 +558,7 @@ void CMemText::MemoryLoad(CHAR * pMem, INT size)
 	while( fwd - pMem < size ) {
 		beg = fwd;
 		while( fwd - pMem < size && * fwd != '\r' && * fwd != '\n' && * fwd != '\0' ) fwd++;
-		AddTail( CString(beg, fwd-beg) );
+		AddTail( CString(beg, (int)(fwd-beg)) );
 		if( * fwd != '\0' ) { if(* fwd == '\r') fwd++; if(* fwd == '\n') fwd++; }
 		else break;
 	}
@@ -729,7 +729,7 @@ BOOL CAnalyzedText::FileSave(LPCTSTR lpszPathName, INT nEncodingType, INT nFileF
 				}
 
 				nCount = MultiByteToWideChar(CP_ACP, 0, rLine, -1, (LPWSTR)pWideBuffer, nBufferSize + 1);
-				INT nLength = wcslen( (LPCWSTR)pWideBuffer );
+				INT nLength = (INT)wcslen( (LPCWSTR)pWideBuffer );
 
 				file.Write( pWideBuffer, 2 * nLength );
 				if( pos ) file.Write( szWideDelim, 2 * nDelimSize );
@@ -751,7 +751,7 @@ BOOL CAnalyzedText::FileSave(LPCTSTR lpszPathName, INT nEncodingType, INT nFileF
 				}
 
 				nCount = MultiByteToWideChar(CP_ACP, 0, rLine, -1, (LPWSTR)pWideBuffer, nBufferSize + 1);
-				INT nLength = wcslen( (LPCWSTR)pWideBuffer );
+				INT nLength = (INT)wcslen( (LPCWSTR)pWideBuffer );
 				for( INT i = 0; i < nLength; i++ ) _SWAP_UCHAR( pWideBuffer[2*i], pWideBuffer[2*i+1] );
 
 				file.Write( pWideBuffer, 2 * nLength );
@@ -773,7 +773,7 @@ BOOL CAnalyzedText::FileSave(LPCTSTR lpszPathName, INT nEncodingType, INT nFileF
 
 				nCount = MultiByteToWideChar(CP_ACP, 0, rLine, -1, (LPWSTR)pWideBuffer, nBufferSize + 1);
 				nCount = WideCharToMultiByte(CP_UTF8, 0, (LPCWSTR)pWideBuffer, -1, pBuffer, 3 * (nBufferSize + 1), NULL, NULL);
-				INT nLength = strlen( pBuffer );
+				INT nLength = (INT)strlen( pBuffer );
 
 				file.Write( pBuffer, nLength );
 				if( pos ) file.Write( szDelim, nDelimSize );
@@ -847,7 +847,7 @@ CString CUserCommand::GetHotKeyText()
 {
 	UINT nScanCode = MapVirtualKey( m_wVirtualKeyCode, 0 );
 	LPARAM lParam = nScanCode << 16;
-	TCHAR szKeyName[1024]; GetKeyNameText( lParam, szKeyName, 1024 );
+	TCHAR szKeyName[1024]; GetKeyNameText( (LONG)lParam, szKeyName, 1024 );
 
 	CString szHotKeyText = "";
 	if( strlen(szKeyName) ) {
@@ -955,7 +955,7 @@ CString CMacroBuffer::GetHotKeyText()
 {
 	UINT nScanCode = MapVirtualKey( m_wVirtualKeyCode, 0 );
 	LPARAM lParam = nScanCode << 16;
-	TCHAR szKeyName[1024]; GetKeyNameText( lParam, szKeyName, 1024 );
+	TCHAR szKeyName[1024]; GetKeyNameText( (LONG)lParam, szKeyName, 1024 );
 
 	CString szHotKeyText = "";
 	if( strlen(szKeyName) ) {
@@ -979,28 +979,28 @@ BOOL CMacroBuffer::StreamSave(ofstream & fout)
 	fout.write((const char *)(& nLength), sizeof(nLength));
 	if( nLength ) fout.write((const char *)(LPCTSTR)m_szName, nLength);
 
-	nCount = m_lstAction.GetCount(); pos = m_lstAction.GetHeadPosition();
+	nCount = (INT)m_lstAction.GetCount(); pos = m_lstAction.GetHeadPosition();
 	fout.write((const char *)(& nCount), sizeof(nCount));
 	while( pos ) { 
 		nAction = m_lstAction.GetNext(pos); 
 		fout.write((const char *)(& nAction), sizeof(nAction));
 	}
 
-	nCount = m_lstParam.GetCount(); pos = m_lstParam.GetHeadPosition();
+	nCount = (INT)m_lstParam.GetCount(); pos = m_lstParam.GetHeadPosition();
 	fout.write((const char *)(& nCount), sizeof(nCount));
 	while( pos ) { 
 		nParam = m_lstParam.GetNext(pos); 
 		fout.write((const char *)(& nParam), sizeof(nParam));
 	}
 
-	nCount = m_lstFlags.GetCount(); pos = m_lstFlags.GetHeadPosition();
+	nCount = (INT)m_lstFlags.GetCount(); pos = m_lstFlags.GetHeadPosition();
 	fout.write((const char *)(& nCount), sizeof(nCount));
 	while( pos ) { 
 		nFlags = m_lstFlags.GetNext(pos); 
 		fout.write((const char *)(& nFlags), sizeof(nFlags));
 	}
 
-	nCount = m_lstString.GetCount(); pos = m_lstString.GetHeadPosition();
+	nCount = (INT)m_lstString.GetCount(); pos = m_lstString.GetHeadPosition();
 	fout.write((const char *)(& nCount), sizeof(nCount));
 	while( pos ) { 
 		szString = m_lstString.GetNext(pos); nLength = szString.GetLength();
